Define PortalProps in types
