+++ packages/ui/src/alert-dialog/index.ts
@@ -1,4 +1,4 @@
-export type { PortalProps } from './portal';
+export type { PortalProps } from './types';
 export { Root } from './root';
 export { Trigger } from './trigger';
 export { Popup } from './popup';
