+++ packages/ui/src/form/primitives/select/index.ts
@@ -1,5 +1,5 @@
 export { Item } from './item';
-export type { PortalProps } from './portal';
+export type { PortalProps } from './types';
 export { Portal } from './portal';
 export { Popup } from './popup';
 export { Root } from './root';
