+++ packages/ui/src/popover/types.ts
@@ -1,8 +1,9 @@
-import type { ReactElement, ReactNode } from 'react';
+import type { ComponentPropsWithoutRef, ReactElement, ReactNode } from 'react';
 import type { Popover as _Popover } from '@base-ui/react/popover';
 
 import type { ComponentProps } from '../utils/types';
-import type { PortalProps } from './portal';
+
+export type PortalProps = ComponentPropsWithoutRef< typeof _Popover.Portal >;
 
 export interface RootProps
 	extends Pick<
