+++ packages/ui/src/tooltip/index.ts
@@ -4,5 +4,5 @@ import { Trigger } from './trigger';
 import { Root } from './root';
 import { Provider } from './provider';
 
-export type { PortalProps } from './portal';
+export type { PortalProps } from './types';
 export { Provider, Root, Trigger, Popup, Portal };
